+++ config/routing.py
@@ -1,7 +1,7 @@
 from channels.routing import ProtocolTypeRouter, URLRouter
 from django.core.asgi import get_asgi_application
 from game.routing import websocket_urlpatterns
-from config.middlewares import CustomWsMiddleware
+from config.middleware import CustomWsMiddleware
 
 django_asgi_app = get_asgi_application()
